# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Rayzan Mart API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: categoryId
          in: query
          schema:
            type: string
        - name: brandId
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [price_asc, price_desc, newest, popular]
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"

    post:
      operationId: createProduct
      tags: [products]
      summary: Create product (Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateProduct
      tags: [products]
      summary: Update product (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /products/{id}/reviews:
    get:
      operationId: getProductReviews
      tags: [products]
      summary: Get product reviews
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"

    post:
      operationId: createProductReview
      tags: [products]
      summary: Create a review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # Categories
...
  /me/wishlist:
    get:
      operationId: getWishlist
      tags: [users]
      summary: Get user wishlist
      responses:
        "200":
          description: Wishlist
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WishlistItem"

    post:
      operationId: addToWishlist
      tags: [users]
      summary: Add to wishlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WishlistBody"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishlistItem"
  /me/wishlist/{productId}:
    delete:
      operationId: removeFromWishlist
      tags: [users]
      summary: Remove from wishlist
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # Profile
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
...
    Product:
      type: object
      properties:
        id:
          type: string
        nameBn:
          type: string
        nameEn:
          type: string
        descriptionBn:
          type: string
          nullable: true
        descriptionEn:
          type: string
          nullable: true
        price:
          type: number
        originalPrice:
          type: number
          nullable: true
        imageUrl:
          type: string
          nullable: true
        galleryImages:
          type: array
          items:
            type: string
        categoryId:
          type: string
          nullable: true
        brandId:
          type: string
          nullable: true
        brand:
          type: string
          nullable: true
        stock:
          type: integer
        rating:
          type: number
        reviewsCount:
          type: integer
        isFeatured:
          type: boolean
        discountPercent:
          type: number
        hasVariants:
          type: boolean
        isActive:
          type: boolean
        sku:
          type: string
          nullable: true
        productStatus:
          type: string
        costPrice:
          type: number
          nullable: true
        isAffiliate:
          type: boolean
        createdAt:
          type: string
      required: [id, nameBn, nameEn, price]
    ProductDetail:
      allOf:
        - $ref: "#/components/schemas/Product"
        - type: object
          properties:
            category:
              $ref: "#/components/schemas/Category"
            brandDetail:
              $ref: "#/components/schemas/Brand"
            variants:
              type: array
              items:
                $ref: "#/components/schemas/ProductVariant"
            reviews:
              type: array
              items:
                $ref: "#/components/schemas/Review"
    ProductVariant:
      type: object
      properties:
        id:
          type: string
        productId:
          type: string
        nameEn:
          type: string
        nameBn:
          type: string
        sku:
          type: string
          nullable: true
        price:
          type: number
          nullable: true
        stock:
          type: integer
        attributes:
          type: object
          additionalProperties: true
        imageUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
      required: [id, productId, nameEn, nameBn]
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [products, total, page, limit, totalPages]
    CreateProductBody:
      type: object
      properties:
        nameBn:
          type: string
        nameEn:
          type: string
        descriptionBn:
          type: string
        descriptionEn:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        imageUrl:
          type: string
        galleryImages:
          type: array
          items:
            type: string
        categoryId:
          type: string
        brandId:
          type: string
        stock:
          type: integer
        isFeatured:
          type: boolean
        sku:
          type: string
        costPrice:
          type: number
        isAffiliate:
          type: boolean
        affiliateCommissionType:
          type: string
        affiliateCommissionValue:
          type: number
        hasVariants:
          type: boolean
        productStatus:
          type: string
      required: [nameBn, nameEn, price]
    Category:
      type: object
      properties:
        id:
          type: string
        nameBn:
          type: string
        nameEn:
          type: string
        parentId:
          type: string
          nullable: true
        slug:
          type: string
          nullable: true
        icon:
          type: string
          nullable: true
        sortOrder:
          type: integer
        isActive:
          type: boolean
        productCount:
          type: integer
      required: [id, nameBn, nameEn]
...
    Brand:
      type: object
      properties:
        id:
          type: string
        nameBn:
          type: string
        nameEn:
          type: string
        slug:
          type: string
          nullable: true
        logoUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
        productCount:
          type: integer
      required: [id, nameBn, nameEn]
...
    Review:
      type: object
      properties:
        id:
          type: string
        productId:
          type: string
        userId:
          type: string
          nullable: true
        orderId:
          type: string
          nullable: true
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
          nullable: true
        isApproved:
          type: boolean
        createdAt:
          type: string
      required: [id, productId, rating]
    CreateReviewBody:
      type: object
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
        orderId:
          type: string
      required: [rating]
    WishlistItem:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        productId:
          type: string
        product:
          $ref: "#/components/schemas/Product"
        createdAt:
          type: string
      required: [id, userId, productId]
    WishlistBody:
      type: object
      properties:
        productId:
          type: string
      required: [productId]
...