# lib/api-zod/src/generated/types/createProductBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rayzan Mart API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProductBody {
  nameBn: string;
  nameEn: string;
  descriptionBn?: string;
  descriptionEn?: string;
  price: number;
  originalPrice?: number;
  imageUrl?: string;
  galleryImages?: string[];
  categoryId?: string;
  brandId?: string;
  stock?: number;
  isFeatured?: boolean;
  sku?: string;
  costPrice?: number;
  isAffiliate?: boolean;
  affiliateCommissionType?: string;
  affiliateCommissionValue?: number;
  hasVariants?: boolean;
  productStatus?: string;
}