# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rayzan Mart API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: string;
  nameBn: string;
  nameEn: string;
  descriptionBn?: string | null;
  descriptionEn?: string | null;
  price: number;
  originalPrice?: number | null;
  imageUrl?: string | null;
  galleryImages?: string[];
  categoryId?: string | null;
  brandId?: string | null;
  brand?: string | null;
  stock?: number;
  rating?: number;
  reviewsCount?: number;
  isFeatured?: boolean;
  discountPercent?: number;
  hasVariants?: boolean;
  isActive?: boolean;
  sku?: string | null;
  productStatus?: string;
  costPrice?: number | null;
  isAffiliate?: boolean;
  createdAt?: string;
}